Fix async compatibility with Internet Explorer 10
This fixes async to work on Internet Explorer 10. Internet Explorer 10
requires that the `setImmediate` function is not called as a method. As
such, it cannot simply be aliased as `async.setImmediate` and
`async.nextTick` as `Invalid calling object` error will be encountered
since `setImmediate` will be called with `this` set to `async`.