Cache length in cross-browser _each() for loop
In the for loop on line 48 the length of the array is being looked up on each iteration of the loop. This change caches the length of the array in a variable to optimize for speed. This way there is only one lookup for array length. Nothing else is changed.
LINE 48 BEFORE: for (var i = 0; i < arr.length; i += 1) {
LINE 48 AFTER: for (var i = 0, l = arr.length; i < l; i += 1) {