Implement contextSeries, alternative to waterfall.
This function is mainly useful as an alternative to watefall in versions of javascript, which support destructing.

Main problem with waterfall is that it becomes hard to debug/refactor after initial write due to needing to pass
around arguments in _exactly_ the correct order. This function when used with destructing gets around that limitation
by allowing to name steps in the series and then each function can grab results it cares about.