Use nexttick instead of setImmediate
Basically this change is based on @isaacs reflexion on this post:

http://blog.izs.me/post/59142742143/designing-apis-for-asynchrony

```
This uses Node’s synthetic deferral function: process.nextTick, which runs the supplied callback at the end of the current run-to-completion. You can also use setImmediate, but that’s slightly slower. (Yes, yes, it’s named badly; setImmediate is slightly less “immediate” and more “next” than nextTick, but this is an accident of history we cannot correct without a time machine.)
```