fix: handle ES2017 Async Functions in async.apply()
Previously, `async.apply()` always returns a regular function, even if
an async function was passed in. This meant, for example, that
`async.apply()` would not be able to be used in other methods without
wrapping in `async.asyncify()`.

This fix checks what type of function was passed in, and returns the
same type.