fix: address edge case in comment stripping (#1780)
Unforunately, cdfb4917e6028c8f966276d6e792018c7fd2ae3c introduced issues
with certain edge cases for comments. Fixing the edge case in the
regular expression is likely to reintroduce the issue that commit was
addressing. This change replaces the regular expression entirely with
a function that iterates through the string instead of regular
expressions.